+++ src/SourceAdapters/FileAdapter.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 namespace Plank\Mediable\SourceAdapters;
 
 use Plank\Mediable\Helpers\File as FileHelper;
+use Plank\Mediable\Stream;
+use Psr\Http\Message\StreamInterface;
 use Symfony\Component\HttpFoundation\File\File;
 
 /**
@@ -77,9 +79,9 @@ class FileAdapter implements SourceAdapterInterface
     /**
      * @inheritdoc
      */
-    public function getStreamResource()
+    public function getStream(): ?StreamInterface
     {
-        return fopen($this->path(), 'rb');
+        return new Stream(fopen($this->path(), 'rb'));
     }
 
     /**