+++ src/SourceAdapters/LocalPathAdapter.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 namespace Plank\Mediable\SourceAdapters;
 
 use Plank\Mediable\Helpers\File;
+use Plank\Mediable\Stream;
+use Psr\Http\Message\StreamInterface;
 
 /**
  * Local Path Adapter.
@@ -71,11 +73,11 @@ class LocalPathAdapter implements SourceAdapterInterface
     }
 
     /**
-     * @inheritdoc
+     * {@inheritdoc}
      */
-    public function getStreamResource()
+    public function getStream(): ?StreamInterface
     {
-        return fopen($this->path(), 'rb');
+        return new Stream(fopen($this->path(), 'rb'));
     }
 
     /**